Move state machine related files to `stateMachine` directory
