+++ src/stateMachine/stateHandlers/BaseStateHandler.ts
@@ -1,8 +1,8 @@
-import type { IntermediateState } from '../typings/IntermediateState';
-import type { TerminalState } from '../typings/TerminalState';
-import type { JSONValue } from '../typings/JSONValue';
-import type { BaseState } from '../typings/BaseState';
-import type { ExecutionResult } from '../typings/StateHandlers';
+import type { IntermediateState } from '../../typings/IntermediateState';
+import type { TerminalState } from '../../typings/TerminalState';
+import type { JSONValue } from '../../typings/JSONValue';
+import type { BaseState } from '../../typings/BaseState';
+import type { ExecutionResult } from '../../typings/StateHandlers';
 
 abstract class BaseStateHandler<T extends BaseState | IntermediateState | TerminalState> {
   protected stateDefinition: T;
